refactor(core): organize classes into subpackages (console, repl, snapshot, publish) (#218)
- Move ConsoleOutputListener/ConsoleController to core.console
- Move ReplManager to core.repl
- Move NormalizedSnapshot/ReplSnapshotAdapter/DebugSnapshotAdapter to core.snapshot
- Move PayloadPublisher to core.publish
- Update imports across server, ui, core, and tests
- No functional changes; build/tests pass
