+++ src/main/java/de/code14/edupydebugger/core/snapshot/NormalizedSnapshot.java
@@ -1,4 +1,4 @@
-package de.code14.edupydebugger.core;
+package de.code14.edupydebugger.core.snapshot;
 
 import de.code14.edupydebugger.analysis.dynamicanalysis.ObjectInfo;
 import de.code14.edupydebugger.server.dto.VariableDTO;
@@ -6,10 +6,7 @@ import de.code14.edupydebugger.server.dto.VariableDTO;
 import java.util.List;
 import java.util.Map;
 
-/**
- * Normalized view of variables and objects for publishing to the frontend,
- * independent of the data source (Debugger vs. REPL).
- */
+/** Normalized view of variables and objects for publishing. */
 public record NormalizedSnapshot(List<VariableDTO> variables,
                                  Map<String, ObjectInfo> objects) {}
