+++ src/main/java/de/code14/edupydebugger/server/DebugServerEndpoint.java
@@ -5,9 +5,9 @@ import com.google.gson.JsonSyntaxException;
 import com.intellij.execution.process.ProcessHandler;
 import com.intellij.openapi.diagnostic.Logger;
 import com.jetbrains.python.debugger.PyDebugProcess;
-import de.code14.edupydebugger.core.ConsoleController;
-import de.code14.edupydebugger.core.ConsoleOutputListener;
-import de.code14.edupydebugger.core.ReplManager;
+import de.code14.edupydebugger.core.console.ConsoleController;
+import de.code14.edupydebugger.core.console.ConsoleOutputListener;
+import de.code14.edupydebugger.core.repl.ReplManager;
 import de.code14.edupydebugger.core.DebugProcessController;
 import de.code14.edupydebugger.core.DebugSessionController;
 import de.code14.edupydebugger.server.dto.*;
